Enabled cscoperelative (csre) and removed cspc.
*cscoperelative* *csre*
If 'cscoperelative' is set, then in absence of a prefix given to cscope
(prefix is the argument of -P option of cscope), basename of cscope.out
location (usually the project root directory) will be used as the prefix
to construct an absolute path. The default is off.

Also removed cspc option to see full file path

*cscopepathcomp* *cspc*
The value of 'cspc' determines how many components of a file's path to
display.